+++ src/components/App/components/index.js
@@ -1,0 +1,2 @@
+export * from './Navbar';
+export * from './PrivateRoute';
