--- src/components/index.js
@@ -1,3 +1,0 @@
-export { Navbar } from "./Navbar";
-export { VideoCard } from "./VideoCard";
-export { Dropdown } from "./Dropdown";
