+++ src/index.js
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 import DataContextProvider from "./contexts/DataContextProvider";
-import App from "./App";
+import App from "./components/App/App";
 import { BrowserRouter as Router } from "react-router-dom";
 
 const rootElement = document.getElementById("root");
