--- src/pages/index.js
@@ -1,6 +1,0 @@
-export { Home } from "./Home";
-export { Playlist } from "./Playlist";
-export { VideoPlayer } from "./VideoPlayer";
-export { PlaylistListing } from "./PlaylistListing";
-export { Account } from "./Account";
-export { NotFound } from "./NotFound";
